requirements.txt
blurhash==1.1.4
certifi==2021.5.30
chardet==4.0.0
decorator==5.0.9
idna==2.10
Mastodon.py==1.5.1
oauthlib==3.1.1
PySocks==1.7.1
python-dateutil==2.8.1
python-magic==0.4.24
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.16.0
tweepy==3.10.0
urllib3==1.26.5
